bot/requirements.txt
google-cloud-storage==3.9.0
libmozdata==0.2.12
pytz==2026.1.post1
pyyaml==6.0.3
sentry-sdk==2.54.0
structlog==25.5.0
taskcluster==97.0.1
tenacity==9.1.4
tqdm==4.67.3
zstandard==0.25.0
